Refactor trash handling
Use Desktop.moveToTrash() instead of platform-specific methods. This
further reduces on native methods of SystemUtilities, pushing us further
from moving away from native SystemUtilities functions in general and
therefore allowing us to drop the pre-compiled native binaries.

Note that moving files to the trash is still unsupported on Linux. This
is due to a limitation with Desktop.moveToTrash() apparently not being
supported on Linux. The JDK will have to update this method so Linux is
properly supported. Until then, only Windows and macOS are supported, but
now they use a common implementation rather than platform-specific
implementations.